+++ tests/sdk/test_runner.py
@@ -53,6 +53,7 @@ def test_apply_cli_options(mocker, cli):
         '--socket', '@katana-127-0-0-1-5010-foo',
         '--tcp', '5010',
         '--debug',
+        '--action', 'foo_action',
         '--disable-compact-names',
         '--quiet',
         '--var', 'foo=bar',
@@ -73,6 +74,7 @@ def test_apply_cli_options(mocker, cli):
         'socket': '@katana-127-0-0-1-5010-foo',
         'tcp': 5010,
         'debug': True,
+        'action': 'foo_action',
         'disable_compact_names': True,
         'quiet': True,
         'var': {'foo': 'bar', 'hello': 'world'},
@@ -176,11 +178,12 @@ def test_component_run(mocker, cli):
 
     # Check server creation arguments
     args, kwargs = ServerCls.call_args
-    assert args == ('ipc://{}'.format(socket), callbacks, {
+    assert args == (callbacks, {
         'name': 'foo',
         'version': '1.0',
         'component': 'service',
         'framework_version': '1.0.0',
+        'action': None,
         'socket': socket,
         'tcp': None,
         'debug': True,
@@ -195,7 +198,8 @@ def test_component_run(mocker, cli):
     assert kwargs.get('error_callback') == error_callback
 
     # Check that server was run using greenlets
-    gevent_spawn.assert_called_once_with(server.listen)
+    channel = 'ipc://@katana-127-0-0-1-5010-foo'
+    gevent_spawn.assert_called_once_with(server.listen, channel)
     gevent_signal.assert_called()
     greenlet.join.assert_called()